refactor newlines offset approach for reusability
